LaekningGenAIRecommendationModule/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
openai
azure-identity
azure-keyvault-secrets
pyodbc